+++ strawberry/http.py
@@ -1,9 +1,11 @@
+from dataclasses import dataclass
 from typing import Any, Dict, List, Optional
 
 from typing_extensions import TypedDict
 
 from graphql.error import format_error as format_graphql_error
 
+from strawberry.exceptions import MissingQueryError
 from strawberry.types import ExecutionResult
 
 
@@ -19,3 +21,23 @@ def process_result(result: ExecutionResult) -> GraphQLHTTPResponse:
         data["errors"] = [format_graphql_error(err) for err in result.errors]
 
     return data
+
+
+@dataclass
+class GraphQLRequestData:
+    query: str
+    variables: Optional[Dict[str, Any]]
+    operation_name: Optional[str]
+
+
+def parse_request_data(data: Dict) -> GraphQLRequestData:
+    if "query" not in data:
+        raise MissingQueryError()
+
+    result = GraphQLRequestData(
+        query=data["query"],
+        variables=data.get("variables"),
+        operation_name=data.get("operationName"),
+    )
+
+    return result